new: support for blog posts and refactors
To make it easier to convert the files to HTML I made a general
function, which takes in the file ($1) and an optional directory ($2)
with that I then implemented the blog pages.

The generation of the entries.html needs to be configured a little bit,
right now it's kinda horrid, and I also need to add dates.